+++ fluorender/FluoRender/Formats/oif_reader.cpp
@@ -85,13 +85,13 @@ void OIFReader::Preprocess()
    m_oif_info.clear();
 
    //separate path and name
-   size_t pos = m_path_name.find_last_of(GETSLASH());
+   int64_t pos = m_path_name.find_last_of(GETSLASH());
    if (pos == -1)
       return;
    wstring path = m_path_name.substr(0, pos+1);
    wstring name = m_path_name.substr(pos+1);
    //extract time sequence string
-   size_t begin = name.find(m_time_id);
+   int64_t begin = name.find(m_time_id);
    size_t end = -1;
    size_t id_len = m_time_id.size();
    if (begin != -1)
@@ -121,7 +121,7 @@ void OIFReader::Preprocess()
       std::vector<std::wstring> list;
       int tmp = 0;
       FIND_FILES(path,L".oif",list,tmp,name.substr(0,begin+id_len+1));
-      for(int i = 0; i < list.size(); i++) {
+      for(size_t i = 0; i < list.size(); i++) {
          size_t start_idx = list.at(i).find(m_time_id) + id_len;
          size_t end_idx   = list.at(i).find(L".oif");
          size_t size = end_idx - start_idx;
@@ -244,8 +244,8 @@ void OIFReader::ReadTifSequence(wstring file_name)
    if (file_name.substr(line_size-3, 3) == L"tif")
    {
       //interpret file_name
-      size_t pos;
-      size_t pos_ = file_name.find_last_of(L'_');
+      int64_t pos;
+      int64_t pos_ = file_name.find_last_of(L'_');
       if (pos_!=-1)
       {
          size_t j;
@@ -531,19 +531,19 @@ void OIFReader::ReadOifLine(wstring oneline)
          spc= fabs((WSTOD(end_pos)-
                   WSTOD(start_pos)))/
             dmax;
-      if (pix_unit.find(L"nm")!=-1)
+      if ((int64_t)pix_unit.find(L"nm")!=-1)
          spc /= 1000.0;
-      if (axis_code.find(L"X")!=-1)
+      if ((int64_t)axis_code.find(L"X")!=-1)
       {
          m_x_size = WSTOI(max_size);
          m_xspc = spc;
       }
-      else if (axis_code.find(L"Y")!=-1)
+      else if ((int64_t)axis_code.find(L"Y")!=-1)
       {
          m_y_size = WSTOI(max_size);
          m_yspc = spc;
       }
-      else if (axis_code.find(L"Z")!=-1)
+      else if ((int64_t)axis_code.find(L"Z")!=-1)
          m_zspc = spc;
 
       axis_code.clear();
